Workaround to inject CSS | onLoad package is now separate
