+++ src/plugins/fetch-plugin.ts
@@ -1,12 +1,6 @@
 import * as esbuild from "esbuild-wasm";
-import axios from "axios";
-import localForage from "localforage";
 
-const bifrostCache = localForage.createInstance({
-	name: "bifrost",
-});
-
-export const fetchPlugin = (codeBlock: string) => {
+export const fetchPlugin = () => {
 	return {
 		name: "unpkg-path-plugin",
 		setup(build: esbuild.PluginBuild) {
@@ -39,35 +33,6 @@ export const fetchPlugin = (codeBlock: string) => {
 					namespace: "a",
 					path: `https://unpkg.com/${args.path}`,
 				};
-			});
-
-			/**
-			 *
-			 */
-
-			build.onLoad({ filter: /.*/ }, async (args: any) => {
-				if (args.path === "index.js") {
-					return {
-						loader: "jsx",
-						contents: codeBlock,
-					};
-				}
-
-				const cachedResult = await bifrostCache.getItem<esbuild.OnLoadResult>(
-					args.path
-				);
-
-				if (cachedResult) {
-					return cachedResult;
-				}
-				const { data, request } = await axios.get(args.path);
-				const result: esbuild.OnLoadResult = {
-					loader: "jsx",
-					contents: data,
-					resolveDir: new URL("./", request.responseURL).pathname,
-				};
-				await bifrostCache.setItem(args.path, result);
-				return result;
 			});
 		},
 	};
